pyproject.toml
[project]
name = "end-to-end-rag"
version = "1.0.0"
description = "端到端 RAG 評估框架，支援多種檢索方法與全面的評估指標"
authors = [
    {name = "End_to_End_RAG Team"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["RAG", "retrieval", "evaluation", "LLM", "graph-rag", "vector-search"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    # Core Dependencies
    "llama-index-core>=0.10.0",
    "llama-index-llms-ollama>=0.1.0",
    "llama-index-embeddings-ollama>=0.1.0",
    
    # RAG Frameworks
    "lightrag>=0.1.0",
    
    # Evaluation Metrics
    "evaluate>=0.4.0",
    "rouge-score>=0.1.0",
    "bert-score>=0.3.0",
    "sacrebleu>=2.0.0",
    "nltk>=3.8.0",
    "jieba>=0.42.0",
    
    # Data Processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0.0",
    
    # Utilities
    "tqdm>=4.65.0",
    "nest-asyncio>=1.5.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    
    # Code Quality
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
    
    # Documentation
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/End_to_End_RAG"
Documentation = "https://github.com/yourusername/End_to_End_RAG/blob/main/README.md"
Repository = "https://github.com/yourusername/End_to_End_RAG"
Issues = "https://github.com/yourusername/End_to_End_RAG/issues"

[project.scripts]
rag-evaluate = "scripts.run_evaluation:main"
rag-test = "scripts.run_comprehensive_tests:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src", "src.config", "src.data", "src.rag", "src.rag.wrappers", 
            "src.rag.vector", "src.rag.graph", "src.rag.schema", "src.rag.retrieval",
            "src.graph_builder", "src.graph_retriever", "src.evaluation", 
            "src.evaluation.metrics", "src.plugins", "src.storage", "src.utils"]

[tool.setuptools.package-data]
"*" = ["*.yml", "*.yaml", "*.json"]

# ============================================================================
# Testing Configuration
# ============================================================================

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto"

# ============================================================================
# Code Formatting Configuration
# ============================================================================

[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | legacy
  | storage
  | results
)/
'''

# ============================================================================
# Import Sorting Configuration
# ============================================================================

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["legacy/*", "storage/*", "results/*"]

# ============================================================================
# Type Checking Configuration
# ============================================================================

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # 逐步啟用
ignore_missing_imports = true
exclude = [
    'legacy',
    'storage',
    'results',
    'build',
]

[[tool.mypy.overrides]]
module = [
    "lightrag.*",
    "llama_index.*",
    "bert_score.*",
    "rouge_score.*",
]
ignore_missing_imports = true

# ============================================================================
# Coverage Configuration
# ============================================================================

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/legacy/*",
    "*/__pycache__/*",
    "*/build/*",
    "*/dist/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# ============================================================================
# Flake8 Configuration (in setup.cfg or .flake8 file)
# ============================================================================

# Note: flake8 doesn't support pyproject.toml yet
# Create a .flake8 file with these settings:
#
# [flake8]
# max-line-length = 100
# extend-ignore = E203, E266, E501, W503
# exclude = 
#     .git,
#     __pycache__,
#     legacy,
#     storage,
#     results,
#     build,
#     dist
# max-complexity = 10
